requirements.txt
mysql-connector-python==9.1.0
streamlit==1.40.2
pathlib
pdfplumber==0.11.4
pytesseract==0.3.13
pillow==11.0.0
pdf2image==1.17.0
docx2txt==0.8
google-generativeai==0.8.3
